Enum aws_sdk_codedeploy::Error

#[non_exhaustive]
pub enum Error {

    AlarmsLimitExceededException(AlarmsLimitExceededException),
    ApplicationAlreadyExistsException(ApplicationAlreadyExistsException),
    ApplicationDoesNotExistException(ApplicationDoesNotExistException),
    ApplicationLimitExceededException(ApplicationLimitExceededException),
    ApplicationNameRequiredException(ApplicationNameRequiredException),
    ArnNotSupportedException(ArnNotSupportedException),
    BatchLimitExceededException(BatchLimitExceededException),
    BucketNameFilterRequiredException(BucketNameFilterRequiredException),
    DeploymentAlreadyCompletedException(DeploymentAlreadyCompletedException),
    DeploymentConfigAlreadyExistsException(DeploymentConfigAlreadyExistsException),
    DeploymentConfigDoesNotExistException(DeploymentConfigDoesNotExistException),
    DeploymentConfigInUseException(DeploymentConfigInUseException),
    DeploymentConfigLimitExceededException(DeploymentConfigLimitExceededException),
    DeploymentConfigNameRequiredException(DeploymentConfigNameRequiredException),
    DeploymentDoesNotExistException(DeploymentDoesNotExistException),
    DeploymentGroupAlreadyExistsException(DeploymentGroupAlreadyExistsException),
    DeploymentGroupDoesNotExistException(DeploymentGroupDoesNotExistException),
    DeploymentGroupLimitExceededException(DeploymentGroupLimitExceededException),
    DeploymentGroupNameRequiredException(DeploymentGroupNameRequiredException),
    DeploymentIdRequiredException(DeploymentIdRequiredException),
    DeploymentIsNotInReadyStateException(DeploymentIsNotInReadyStateException),
    DeploymentLimitExceededException(DeploymentLimitExceededException),
    DeploymentNotStartedException(DeploymentNotStartedException),
    DeploymentTargetDoesNotExistException(DeploymentTargetDoesNotExistException),
    DeploymentTargetIdRequiredException(DeploymentTargetIdRequiredException),
    DeploymentTargetListSizeExceededException(DeploymentTargetListSizeExceededException),
    DescriptionTooLongException(DescriptionTooLongException),
    EcsServiceMappingLimitExceededException(EcsServiceMappingLimitExceededException),
    GitHubAccountTokenDoesNotExistException(GitHubAccountTokenDoesNotExistException),
    GitHubAccountTokenNameRequiredException(GitHubAccountTokenNameRequiredException),
    IamArnRequiredException(IamArnRequiredException),
    IamSessionArnAlreadyRegisteredException(IamSessionArnAlreadyRegisteredException),
    IamUserArnAlreadyRegisteredException(IamUserArnAlreadyRegisteredException),
    IamUserArnRequiredException(IamUserArnRequiredException),
    InstanceDoesNotExistException(InstanceDoesNotExistException),
    InstanceIdRequiredException(InstanceIdRequiredException),
    InstanceLimitExceededException(InstanceLimitExceededException),
    InstanceNameAlreadyRegisteredException(InstanceNameAlreadyRegisteredException),
    InstanceNameRequiredException(InstanceNameRequiredException),
    InstanceNotRegisteredException(InstanceNotRegisteredException),
    InvalidAlarmConfigException(InvalidAlarmConfigException),
    InvalidApplicationNameException(InvalidApplicationNameException),
    InvalidArnException(InvalidArnException),
    InvalidAutoRollbackConfigException(InvalidAutoRollbackConfigException),
    InvalidAutoScalingGroupException(InvalidAutoScalingGroupException),
    InvalidBlueGreenDeploymentConfigurationException(InvalidBlueGreenDeploymentConfigurationException),
    InvalidBucketNameFilterException(InvalidBucketNameFilterException),
    InvalidComputePlatformException(InvalidComputePlatformException),
    InvalidDeployedStateFilterException(InvalidDeployedStateFilterException),
    InvalidDeploymentConfigNameException(InvalidDeploymentConfigNameException),
    InvalidDeploymentGroupNameException(InvalidDeploymentGroupNameException),
    InvalidDeploymentIdException(InvalidDeploymentIdException),
    InvalidDeploymentInstanceTypeException(InvalidDeploymentInstanceTypeException),
    InvalidDeploymentStatusException(InvalidDeploymentStatusException),
    InvalidDeploymentStyleException(InvalidDeploymentStyleException),
    InvalidDeploymentTargetIdException(InvalidDeploymentTargetIdException),
    InvalidDeploymentWaitTypeException(InvalidDeploymentWaitTypeException),
    InvalidEc2TagCombinationException(InvalidEc2TagCombinationException),
    InvalidEc2TagException(InvalidEc2TagException),
    InvalidEcsServiceException(InvalidEcsServiceException),
    InvalidExternalIdException(InvalidExternalIdException),
    InvalidFileExistsBehaviorException(InvalidFileExistsBehaviorException),
    InvalidGitHubAccountTokenException(InvalidGitHubAccountTokenException),
    InvalidGitHubAccountTokenNameException(InvalidGitHubAccountTokenNameException),
    InvalidIamSessionArnException(InvalidIamSessionArnException),
    InvalidIamUserArnException(InvalidIamUserArnException),
    InvalidIgnoreApplicationStopFailuresValueException(InvalidIgnoreApplicationStopFailuresValueException),
    InvalidInputException(InvalidInputException),
    InvalidInstanceNameException(InvalidInstanceNameException),
    InvalidInstanceStatusException(InvalidInstanceStatusException),
    InvalidInstanceTypeException(InvalidInstanceTypeException),
    InvalidKeyPrefixFilterException(InvalidKeyPrefixFilterException),
    InvalidLifecycleEventHookExecutionIdException(InvalidLifecycleEventHookExecutionIdException),
    InvalidLifecycleEventHookExecutionStatusException(InvalidLifecycleEventHookExecutionStatusException),
    InvalidLoadBalancerInfoException(InvalidLoadBalancerInfoException),
    InvalidMinimumHealthyHostValueException(InvalidMinimumHealthyHostValueException),
    InvalidNextTokenException(InvalidNextTokenException),
    InvalidOnPremisesTagCombinationException(InvalidOnPremisesTagCombinationException),
    InvalidOperationException(InvalidOperationException),
    InvalidRegistrationStatusException(InvalidRegistrationStatusException),
    InvalidRevisionException(InvalidRevisionException),
    InvalidRoleException(InvalidRoleException),
    InvalidSortByException(InvalidSortByException),
    InvalidSortOrderException(InvalidSortOrderException),
    InvalidTagException(InvalidTagException),
    InvalidTagFilterException(InvalidTagFilterException),
    InvalidTagsToAddException(InvalidTagsToAddException),
    InvalidTargetFilterNameException(InvalidTargetFilterNameException),
    InvalidTargetGroupPairException(InvalidTargetGroupPairException),
    InvalidTargetInstancesException(InvalidTargetInstancesException),
    InvalidTimeRangeException(InvalidTimeRangeException),
    InvalidTrafficRoutingConfigurationException(InvalidTrafficRoutingConfigurationException),
    InvalidTriggerConfigException(InvalidTriggerConfigException),
    InvalidUpdateOutdatedInstancesOnlyValueException(InvalidUpdateOutdatedInstancesOnlyValueException),
    LifecycleEventAlreadyCompletedException(LifecycleEventAlreadyCompletedException),
    LifecycleHookLimitExceededException(LifecycleHookLimitExceededException),
    MultipleIamArnsProvidedException(MultipleIamArnsProvidedException),
    OperationNotSupportedException(OperationNotSupportedException),
    ResourceArnRequiredException(ResourceArnRequiredException),
    ResourceValidationException(ResourceValidationException),
    RevisionDoesNotExistException(RevisionDoesNotExistException),
    RevisionRequiredException(RevisionRequiredException),
    RoleRequiredException(RoleRequiredException),
    TagLimitExceededException(TagLimitExceededException),
    TagRequiredException(TagRequiredException),
    TagSetListLimitExceededException(TagSetListLimitExceededException),
    ThrottlingException(ThrottlingException),
    TriggerTargetsLimitExceededException(TriggerTargetsLimitExceededException),
    UnsupportedActionForDeploymentTypeException(UnsupportedActionForDeploymentTypeException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlarmsLimitExceededException(AlarmsLimitExceededException)

The maximum number of alarms for a deployment group (10) was exceeded.

Tuple Fields of AlarmsLimitExceededException

0: AlarmsLimitExceededException

ApplicationAlreadyExistsException(ApplicationAlreadyExistsException)

An application with the specified name with the IAM user or AWS account already exists.

Tuple Fields of ApplicationAlreadyExistsException

0: ApplicationAlreadyExistsException

ApplicationDoesNotExistException(ApplicationDoesNotExistException)

The application does not exist with the IAM user or AWS account.

Tuple Fields of ApplicationDoesNotExistException

0: ApplicationDoesNotExistException

ApplicationLimitExceededException(ApplicationLimitExceededException)

More applications were attempted to be created than are allowed.

Tuple Fields of ApplicationLimitExceededException

0: ApplicationLimitExceededException

ApplicationNameRequiredException(ApplicationNameRequiredException)

The minimum number of required application names was not specified.

Tuple Fields of ApplicationNameRequiredException

0: ApplicationNameRequiredException

ArnNotSupportedException(ArnNotSupportedException)

The specified ARN is not supported. For example, it might be an ARN for a resource that is not expected.

Tuple Fields of ArnNotSupportedException

0: ArnNotSupportedException

BatchLimitExceededException(BatchLimitExceededException)

The maximum number of names or IDs allowed for this request (100) was exceeded.

Tuple Fields of BatchLimitExceededException

0: BatchLimitExceededException

BucketNameFilterRequiredException(BucketNameFilterRequiredException)

A bucket name is required, but was not provided.

Tuple Fields of BucketNameFilterRequiredException

0: BucketNameFilterRequiredException

DeploymentAlreadyCompletedException(DeploymentAlreadyCompletedException)

The deployment is already complete.

Tuple Fields of DeploymentAlreadyCompletedException

0: DeploymentAlreadyCompletedException

DeploymentConfigAlreadyExistsException(DeploymentConfigAlreadyExistsException)

A deployment configuration with the specified name with the IAM user or AWS account already exists.

Tuple Fields of DeploymentConfigAlreadyExistsException

0: DeploymentConfigAlreadyExistsException

DeploymentConfigDoesNotExistException(DeploymentConfigDoesNotExistException)

The deployment configuration does not exist with the IAM user or AWS account.

Tuple Fields of DeploymentConfigDoesNotExistException

0: DeploymentConfigDoesNotExistException

DeploymentConfigInUseException(DeploymentConfigInUseException)

The deployment configuration is still in use.

Tuple Fields of DeploymentConfigInUseException

0: DeploymentConfigInUseException

DeploymentConfigLimitExceededException(DeploymentConfigLimitExceededException)

The deployment configurations limit was exceeded.

Tuple Fields of DeploymentConfigLimitExceededException

0: DeploymentConfigLimitExceededException

DeploymentConfigNameRequiredException(DeploymentConfigNameRequiredException)

The deployment configuration name was not specified.

Tuple Fields of DeploymentConfigNameRequiredException

0: DeploymentConfigNameRequiredException

DeploymentDoesNotExistException(DeploymentDoesNotExistException)

The deployment with the IAM user or AWS account does not exist.

Tuple Fields of DeploymentDoesNotExistException

0: DeploymentDoesNotExistException

DeploymentGroupAlreadyExistsException(DeploymentGroupAlreadyExistsException)

A deployment group with the specified name with the IAM user or AWS account already exists.

Tuple Fields of DeploymentGroupAlreadyExistsException

0: DeploymentGroupAlreadyExistsException

DeploymentGroupDoesNotExistException(DeploymentGroupDoesNotExistException)

The named deployment group with the IAM user or AWS account does not exist.

Tuple Fields of DeploymentGroupDoesNotExistException

0: DeploymentGroupDoesNotExistException

DeploymentGroupLimitExceededException(DeploymentGroupLimitExceededException)

The deployment groups limit was exceeded.

Tuple Fields of DeploymentGroupLimitExceededException

0: DeploymentGroupLimitExceededException

DeploymentGroupNameRequiredException(DeploymentGroupNameRequiredException)

The deployment group name was not specified.

Tuple Fields of DeploymentGroupNameRequiredException

0: DeploymentGroupNameRequiredException

DeploymentIdRequiredException(DeploymentIdRequiredException)

At least one deployment ID must be specified.

Tuple Fields of DeploymentIdRequiredException

0: DeploymentIdRequiredException

DeploymentIsNotInReadyStateException(DeploymentIsNotInReadyStateException)

The deployment does not have a status of Ready and can't continue yet.

Tuple Fields of DeploymentIsNotInReadyStateException

0: DeploymentIsNotInReadyStateException

DeploymentLimitExceededException(DeploymentLimitExceededException)

The number of allowed deployments was exceeded.

Tuple Fields of DeploymentLimitExceededException

0: DeploymentLimitExceededException

DeploymentNotStartedException(DeploymentNotStartedException)

The specified deployment has not started.

Tuple Fields of DeploymentNotStartedException

0: DeploymentNotStartedException

DeploymentTargetDoesNotExistException(DeploymentTargetDoesNotExistException)

The provided target ID does not belong to the attempted deployment.

Tuple Fields of DeploymentTargetDoesNotExistException

0: DeploymentTargetDoesNotExistException

DeploymentTargetIdRequiredException(DeploymentTargetIdRequiredException)

A deployment target ID was not provided.

Tuple Fields of DeploymentTargetIdRequiredException

0: DeploymentTargetIdRequiredException

DeploymentTargetListSizeExceededException(DeploymentTargetListSizeExceededException)

The maximum number of targets that can be associated with an Amazon ECS or AWS Lambda deployment was exceeded. The target list of both types of deployments must have exactly one item. This exception does not apply to EC2/On-premises deployments.

Tuple Fields of DeploymentTargetListSizeExceededException

0: DeploymentTargetListSizeExceededException

DescriptionTooLongException(DescriptionTooLongException)

The description is too long.

Tuple Fields of DescriptionTooLongException

0: DescriptionTooLongException

EcsServiceMappingLimitExceededException(EcsServiceMappingLimitExceededException)

The Amazon ECS service is associated with more than one deployment groups. An Amazon ECS service can be associated with only one deployment group.

Tuple Fields of EcsServiceMappingLimitExceededException

0: EcsServiceMappingLimitExceededException

GitHubAccountTokenDoesNotExistException(GitHubAccountTokenDoesNotExistException)

No GitHub account connection exists with the named specified in the call.

Tuple Fields of GitHubAccountTokenDoesNotExistException

0: GitHubAccountTokenDoesNotExistException

GitHubAccountTokenNameRequiredException(GitHubAccountTokenNameRequiredException)

The call is missing a required GitHub account connection name.

Tuple Fields of GitHubAccountTokenNameRequiredException

0: GitHubAccountTokenNameRequiredException

IamArnRequiredException(IamArnRequiredException)

No IAM ARN was included in the request. You must use an IAM session ARN or IAM user ARN in the request.

Tuple Fields of IamArnRequiredException

0: IamArnRequiredException

IamSessionArnAlreadyRegisteredException(IamSessionArnAlreadyRegisteredException)

The request included an IAM session ARN that has already been used to register a different instance.

Tuple Fields of IamSessionArnAlreadyRegisteredException

0: IamSessionArnAlreadyRegisteredException

IamUserArnAlreadyRegisteredException(IamUserArnAlreadyRegisteredException)

The specified IAM user ARN is already registered with an on-premises instance.

Tuple Fields of IamUserArnAlreadyRegisteredException

0: IamUserArnAlreadyRegisteredException

IamUserArnRequiredException(IamUserArnRequiredException)

An IAM user ARN was not specified.

Tuple Fields of IamUserArnRequiredException

0: IamUserArnRequiredException

InstanceDoesNotExistException(InstanceDoesNotExistException)

The specified instance does not exist in the deployment group.

Tuple Fields of InstanceDoesNotExistException

0: InstanceDoesNotExistException

InstanceIdRequiredException(InstanceIdRequiredException)

The instance ID was not specified.

Tuple Fields of InstanceIdRequiredException

0: InstanceIdRequiredException

InstanceLimitExceededException(InstanceLimitExceededException)

The maximum number of allowed on-premises instances in a single call was exceeded.

Tuple Fields of InstanceLimitExceededException

0: InstanceLimitExceededException

InstanceNameAlreadyRegisteredException(InstanceNameAlreadyRegisteredException)

The specified on-premises instance name is already registered.

Tuple Fields of InstanceNameAlreadyRegisteredException

0: InstanceNameAlreadyRegisteredException

InstanceNameRequiredException(InstanceNameRequiredException)

An on-premises instance name was not specified.

Tuple Fields of InstanceNameRequiredException

0: InstanceNameRequiredException

InstanceNotRegisteredException(InstanceNotRegisteredException)

The specified on-premises instance is not registered.

Tuple Fields of InstanceNotRegisteredException

0: InstanceNotRegisteredException

InvalidAlarmConfigException(InvalidAlarmConfigException)

The format of the alarm configuration is invalid. Possible causes include:

The alarm list is null.

The alarm object is null.

The alarm name is empty or null or exceeds the limit of 255 characters.

Two alarms with the same name have been specified.

The alarm configuration is enabled, but the alarm list is empty.

Tuple Fields of InvalidAlarmConfigException

0: InvalidAlarmConfigException

InvalidApplicationNameException(InvalidApplicationNameException)

The application name was specified in an invalid format.

Tuple Fields of InvalidApplicationNameException

0: InvalidApplicationNameException

InvalidArnException(InvalidArnException)

The specified ARN is not in a valid format.

Tuple Fields of InvalidArnException

0: InvalidArnException

InvalidAutoRollbackConfigException(InvalidAutoRollbackConfigException)

The automatic rollback configuration was specified in an invalid format. For example, automatic rollback is enabled, but an invalid triggering event type or no event types were listed.

Tuple Fields of InvalidAutoRollbackConfigException

0: InvalidAutoRollbackConfigException

InvalidAutoScalingGroupException(InvalidAutoScalingGroupException)

The Auto Scaling group was specified in an invalid format or does not exist.

Tuple Fields of InvalidAutoScalingGroupException

0: InvalidAutoScalingGroupException

InvalidBlueGreenDeploymentConfigurationException(InvalidBlueGreenDeploymentConfigurationException)

The configuration for the blue/green deployment group was provided in an invalid format. For information about deployment configuration format, see CreateDeploymentConfig.

Tuple Fields of InvalidBlueGreenDeploymentConfigurationException

0: InvalidBlueGreenDeploymentConfigurationException

InvalidBucketNameFilterException(InvalidBucketNameFilterException)

The bucket name either doesn't exist or was specified in an invalid format.

Tuple Fields of InvalidBucketNameFilterException

0: InvalidBucketNameFilterException

InvalidComputePlatformException(InvalidComputePlatformException)

The computePlatform is invalid. The computePlatform should be Lambda, Server, or ECS.

Tuple Fields of InvalidComputePlatformException

0: InvalidComputePlatformException

InvalidDeployedStateFilterException(InvalidDeployedStateFilterException)

The deployed state filter was specified in an invalid format.

Tuple Fields of InvalidDeployedStateFilterException

0: InvalidDeployedStateFilterException

InvalidDeploymentConfigNameException(InvalidDeploymentConfigNameException)

The deployment configuration name was specified in an invalid format.

Tuple Fields of InvalidDeploymentConfigNameException

0: InvalidDeploymentConfigNameException

InvalidDeploymentGroupNameException(InvalidDeploymentGroupNameException)

The deployment group name was specified in an invalid format.

Tuple Fields of InvalidDeploymentGroupNameException

0: InvalidDeploymentGroupNameException

InvalidDeploymentIdException(InvalidDeploymentIdException)

At least one of the deployment IDs was specified in an invalid format.

Tuple Fields of InvalidDeploymentIdException

0: InvalidDeploymentIdException

InvalidDeploymentInstanceTypeException(InvalidDeploymentInstanceTypeException)

An instance type was specified for an in-place deployment. Instance types are supported for blue/green deployments only.

Tuple Fields of InvalidDeploymentInstanceTypeException

0: InvalidDeploymentInstanceTypeException

InvalidDeploymentStatusException(InvalidDeploymentStatusException)

The specified deployment status doesn't exist or cannot be determined.

Tuple Fields of InvalidDeploymentStatusException

0: InvalidDeploymentStatusException

InvalidDeploymentStyleException(InvalidDeploymentStyleException)

An invalid deployment style was specified. Valid deployment types include "IN_PLACE" and "BLUE_GREEN." Valid deployment options include "WITH_TRAFFIC_CONTROL" and "WITHOUT_TRAFFIC_CONTROL."

Tuple Fields of InvalidDeploymentStyleException

0: InvalidDeploymentStyleException

InvalidDeploymentTargetIdException(InvalidDeploymentTargetIdException)

The target ID provided was not valid.

Tuple Fields of InvalidDeploymentTargetIdException

0: InvalidDeploymentTargetIdException

InvalidDeploymentWaitTypeException(InvalidDeploymentWaitTypeException)

The wait type is invalid.

Tuple Fields of InvalidDeploymentWaitTypeException

0: InvalidDeploymentWaitTypeException

InvalidEc2TagCombinationException(InvalidEc2TagCombinationException)

A call was submitted that specified both Ec2TagFilters and Ec2TagSet, but only one of these data types can be used in a single call.

Tuple Fields of InvalidEc2TagCombinationException

0: InvalidEc2TagCombinationException

InvalidEc2TagException(InvalidEc2TagException)

The tag was specified in an invalid format.

Tuple Fields of InvalidEc2TagException

0: InvalidEc2TagException

InvalidEcsServiceException(InvalidEcsServiceException)

The Amazon ECS service identifier is not valid.

Tuple Fields of InvalidEcsServiceException

0: InvalidEcsServiceException

InvalidExternalIdException(InvalidExternalIdException)

The external ID was specified in an invalid format.

Tuple Fields of InvalidExternalIdException

0: InvalidExternalIdException

InvalidFileExistsBehaviorException(InvalidFileExistsBehaviorException)

An invalid fileExistsBehavior option was specified to determine how AWS CodeDeploy handles files or directories that already exist in a deployment target location, but weren't part of the previous successful deployment. Valid values include "DISALLOW," "OVERWRITE," and "RETAIN."

Tuple Fields of InvalidFileExistsBehaviorException

0: InvalidFileExistsBehaviorException

InvalidGitHubAccountTokenException(InvalidGitHubAccountTokenException)

The GitHub token is not valid.

Tuple Fields of InvalidGitHubAccountTokenException

0: InvalidGitHubAccountTokenException

InvalidGitHubAccountTokenNameException(InvalidGitHubAccountTokenNameException)

The format of the specified GitHub account connection name is invalid.

Tuple Fields of InvalidGitHubAccountTokenNameException

0: InvalidGitHubAccountTokenNameException

InvalidIamSessionArnException(InvalidIamSessionArnException)

The IAM session ARN was specified in an invalid format.

Tuple Fields of InvalidIamSessionArnException

0: InvalidIamSessionArnException

InvalidIamUserArnException(InvalidIamUserArnException)

The IAM user ARN was specified in an invalid format.

Tuple Fields of InvalidIamUserArnException

0: InvalidIamUserArnException

InvalidIgnoreApplicationStopFailuresValueException(InvalidIgnoreApplicationStopFailuresValueException)

The IgnoreApplicationStopFailures value is invalid. For AWS Lambda deployments, false is expected. For EC2/On-premises deployments, true or false is expected.

Tuple Fields of InvalidIgnoreApplicationStopFailuresValueException

0: InvalidIgnoreApplicationStopFailuresValueException

InvalidInputException(InvalidInputException)

The input was specified in an invalid format.

Tuple Fields of InvalidInputException

0: InvalidInputException

InvalidInstanceNameException(InvalidInstanceNameException)

The on-premises instance name was specified in an invalid format.

Tuple Fields of InvalidInstanceNameException

0: InvalidInstanceNameException

InvalidInstanceStatusException(InvalidInstanceStatusException)

The specified instance status does not exist.

Tuple Fields of InvalidInstanceStatusException

0: InvalidInstanceStatusException

InvalidInstanceTypeException(InvalidInstanceTypeException)

An invalid instance type was specified for instances in a blue/green deployment. Valid values include "Blue" for an original environment and "Green" for a replacement environment.

Tuple Fields of InvalidInstanceTypeException

0: InvalidInstanceTypeException

InvalidKeyPrefixFilterException(InvalidKeyPrefixFilterException)

The specified key prefix filter was specified in an invalid format.

Tuple Fields of InvalidKeyPrefixFilterException

0: InvalidKeyPrefixFilterException

InvalidLifecycleEventHookExecutionIdException(InvalidLifecycleEventHookExecutionIdException)

A lifecycle event hook is invalid. Review the hooks section in your AppSpec file to ensure the lifecycle events and hooks functions are valid.

Tuple Fields of InvalidLifecycleEventHookExecutionIdException

0: InvalidLifecycleEventHookExecutionIdException

InvalidLifecycleEventHookExecutionStatusException(InvalidLifecycleEventHookExecutionStatusException)

The result of a Lambda validation function that verifies a lifecycle event is invalid. It should return Succeeded or Failed.

Tuple Fields of InvalidLifecycleEventHookExecutionStatusException

0: InvalidLifecycleEventHookExecutionStatusException

InvalidLoadBalancerInfoException(InvalidLoadBalancerInfoException)

An invalid load balancer name, or no load balancer name, was specified.

Tuple Fields of InvalidLoadBalancerInfoException

0: InvalidLoadBalancerInfoException

InvalidMinimumHealthyHostValueException(InvalidMinimumHealthyHostValueException)

The minimum healthy instance value was specified in an invalid format.

Tuple Fields of InvalidMinimumHealthyHostValueException

0: InvalidMinimumHealthyHostValueException

InvalidNextTokenException(InvalidNextTokenException)

The next token was specified in an invalid format.

Tuple Fields of InvalidNextTokenException

0: InvalidNextTokenException

InvalidOnPremisesTagCombinationException(InvalidOnPremisesTagCombinationException)

A call was submitted that specified both OnPremisesTagFilters and OnPremisesTagSet, but only one of these data types can be used in a single call.

Tuple Fields of InvalidOnPremisesTagCombinationException

0: InvalidOnPremisesTagCombinationException

InvalidOperationException(InvalidOperationException)

An invalid operation was detected.

Tuple Fields of InvalidOperationException

0: InvalidOperationException

InvalidRegistrationStatusException(InvalidRegistrationStatusException)

The registration status was specified in an invalid format.

Tuple Fields of InvalidRegistrationStatusException

0: InvalidRegistrationStatusException

InvalidRevisionException(InvalidRevisionException)

The revision was specified in an invalid format.

Tuple Fields of InvalidRevisionException

0: InvalidRevisionException

InvalidRoleException(InvalidRoleException)

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group was specified, the specified service role does not grant the appropriate permissions to Amazon EC2 Auto Scaling.

Tuple Fields of InvalidRoleException

0: InvalidRoleException

InvalidSortByException(InvalidSortByException)

The column name to sort by is either not present or was specified in an invalid format.

Tuple Fields of InvalidSortByException

0: InvalidSortByException

InvalidSortOrderException(InvalidSortOrderException)

The sort order was specified in an invalid format.

Tuple Fields of InvalidSortOrderException

0: InvalidSortOrderException

InvalidTagException(InvalidTagException)

The tag was specified in an invalid format.

Tuple Fields of InvalidTagException

0: InvalidTagException

InvalidTagFilterException(InvalidTagFilterException)

The tag filter was specified in an invalid format.

Tuple Fields of InvalidTagFilterException

0: InvalidTagFilterException

InvalidTagsToAddException(InvalidTagsToAddException)

The specified tags are not valid.

Tuple Fields of InvalidTagsToAddException

0: InvalidTagsToAddException

InvalidTargetFilterNameException(InvalidTargetFilterNameException)

The target filter name is invalid.

Tuple Fields of InvalidTargetFilterNameException

0: InvalidTargetFilterNameException

InvalidTargetGroupPairException(InvalidTargetGroupPairException)

A target group pair associated with this deployment is not valid.

Tuple Fields of InvalidTargetGroupPairException

0: InvalidTargetGroupPairException

InvalidTargetInstancesException(InvalidTargetInstancesException)

The target instance configuration is invalid. Possible causes include:

Configuration data for target instances was entered for an in-place deployment.

The limit of 10 tags for a tag type was exceeded.

The combined length of the tag names exceeded the limit.

A specified tag is not currently applied to any instances.

Tuple Fields of InvalidTargetInstancesException

0: InvalidTargetInstancesException

InvalidTimeRangeException(InvalidTimeRangeException)

The specified time range was specified in an invalid format.

Tuple Fields of InvalidTimeRangeException

0: InvalidTimeRangeException

InvalidTrafficRoutingConfigurationException(InvalidTrafficRoutingConfigurationException)

The configuration that specifies how traffic is routed during a deployment is invalid.

Tuple Fields of InvalidTrafficRoutingConfigurationException

0: InvalidTrafficRoutingConfigurationException

InvalidTriggerConfigException(InvalidTriggerConfigException)

The trigger was specified in an invalid format.

Tuple Fields of InvalidTriggerConfigException

0: InvalidTriggerConfigException

InvalidUpdateOutdatedInstancesOnlyValueException(InvalidUpdateOutdatedInstancesOnlyValueException)

The UpdateOutdatedInstancesOnly value is invalid. For AWS Lambda deployments, false is expected. For EC2/On-premises deployments, true or false is expected.

Tuple Fields of InvalidUpdateOutdatedInstancesOnlyValueException

0: InvalidUpdateOutdatedInstancesOnlyValueException

LifecycleEventAlreadyCompletedException(LifecycleEventAlreadyCompletedException)

An attempt to return the status of an already completed lifecycle event occurred.

Tuple Fields of LifecycleEventAlreadyCompletedException

0: LifecycleEventAlreadyCompletedException

LifecycleHookLimitExceededException(LifecycleHookLimitExceededException)

The limit for lifecycle hooks was exceeded.

Tuple Fields of LifecycleHookLimitExceededException

0: LifecycleHookLimitExceededException

MultipleIamArnsProvidedException(MultipleIamArnsProvidedException)

Both an IAM user ARN and an IAM session ARN were included in the request. Use only one ARN type.

Tuple Fields of MultipleIamArnsProvidedException

0: MultipleIamArnsProvidedException

OperationNotSupportedException(OperationNotSupportedException)

The API used does not support the deployment.

Tuple Fields of OperationNotSupportedException

0: OperationNotSupportedException

ResourceArnRequiredException(ResourceArnRequiredException)

The ARN of a resource is required, but was not found.

Tuple Fields of ResourceArnRequiredException

0: ResourceArnRequiredException

ResourceValidationException(ResourceValidationException)

The specified resource could not be validated.

Tuple Fields of ResourceValidationException

0: ResourceValidationException

RevisionDoesNotExistException(RevisionDoesNotExistException)

The named revision does not exist with the IAM user or AWS account.

Tuple Fields of RevisionDoesNotExistException

0: RevisionDoesNotExistException

RevisionRequiredException(RevisionRequiredException)

The revision ID was not specified.

Tuple Fields of RevisionRequiredException

0: RevisionRequiredException

RoleRequiredException(RoleRequiredException)

The role ID was not specified.

Tuple Fields of RoleRequiredException

0: RoleRequiredException

TagLimitExceededException(TagLimitExceededException)

The maximum allowed number of tags was exceeded.

Tuple Fields of TagLimitExceededException

0: TagLimitExceededException

TagRequiredException(TagRequiredException)

A tag was not specified.

Tuple Fields of TagRequiredException

0: TagRequiredException

TagSetListLimitExceededException(TagSetListLimitExceededException)

The number of tag groups included in the tag set list exceeded the maximum allowed limit of 3.

Tuple Fields of TagSetListLimitExceededException

0: TagSetListLimitExceededException

ThrottlingException(ThrottlingException)

An API function was called too frequently.

Tuple Fields of ThrottlingException

0: ThrottlingException

TriggerTargetsLimitExceededException(TriggerTargetsLimitExceededException)

The maximum allowed number of triggers was exceeded.

Tuple Fields of TriggerTargetsLimitExceededException

0: TriggerTargetsLimitExceededException

UnsupportedActionForDeploymentTypeException(UnsupportedActionForDeploymentTypeException)

A call was submitted that is not supported for the specified deployment type.

Tuple Fields of UnsupportedActionForDeploymentTypeException

0: UnsupportedActionForDeploymentTypeException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>