Module aws_sdk_codedeploy::error

Modules

alarms_limit_exceeded_exception

See AlarmsLimitExceededException

application_already_exists_exception

See ApplicationAlreadyExistsException

application_does_not_exist_exception

See ApplicationDoesNotExistException

application_limit_exceeded_exception

See ApplicationLimitExceededException

application_name_required_exception

See ApplicationNameRequiredException

arn_not_supported_exception

See ArnNotSupportedException

batch_limit_exceeded_exception

See BatchLimitExceededException

bucket_name_filter_required_exception

See BucketNameFilterRequiredException

deployment_already_completed_exception

See DeploymentAlreadyCompletedException

deployment_config_already_exists_exception

See DeploymentConfigAlreadyExistsException

deployment_config_does_not_exist_exception

See DeploymentConfigDoesNotExistException

deployment_config_in_use_exception

See DeploymentConfigInUseException

deployment_config_limit_exceeded_exception

See DeploymentConfigLimitExceededException

deployment_config_name_required_exception

See DeploymentConfigNameRequiredException

deployment_does_not_exist_exception

See DeploymentDoesNotExistException

deployment_group_already_exists_exception

See DeploymentGroupAlreadyExistsException

deployment_group_does_not_exist_exception

See DeploymentGroupDoesNotExistException

deployment_group_limit_exceeded_exception

See DeploymentGroupLimitExceededException

deployment_group_name_required_exception

See DeploymentGroupNameRequiredException

deployment_id_required_exception

See DeploymentIdRequiredException

deployment_is_not_in_ready_state_exception

See DeploymentIsNotInReadyStateException

deployment_limit_exceeded_exception

See DeploymentLimitExceededException

deployment_not_started_exception

See DeploymentNotStartedException

deployment_target_does_not_exist_exception

See DeploymentTargetDoesNotExistException

deployment_target_id_required_exception

See DeploymentTargetIdRequiredException

deployment_target_list_size_exceeded_exception

See DeploymentTargetListSizeExceededException

description_too_long_exception

See DescriptionTooLongException

ecs_service_mapping_limit_exceeded_exception

See EcsServiceMappingLimitExceededException

git_hub_account_token_does_not_exist_exception

See GitHubAccountTokenDoesNotExistException

git_hub_account_token_name_required_exception

See GitHubAccountTokenNameRequiredException

iam_arn_required_exception

See IamArnRequiredException

iam_session_arn_already_registered_exception

See IamSessionArnAlreadyRegisteredException

iam_user_arn_already_registered_exception

See IamUserArnAlreadyRegisteredException

iam_user_arn_required_exception

See IamUserArnRequiredException

instance_does_not_exist_exception

See InstanceDoesNotExistException

instance_id_required_exception

See InstanceIdRequiredException

instance_limit_exceeded_exception

See InstanceLimitExceededException

instance_name_already_registered_exception

See InstanceNameAlreadyRegisteredException

instance_name_required_exception

See InstanceNameRequiredException

instance_not_registered_exception

See InstanceNotRegisteredException

invalid_alarm_config_exception

See InvalidAlarmConfigException

invalid_application_name_exception

See InvalidApplicationNameException

invalid_arn_exception

See InvalidArnException

invalid_auto_rollback_config_exception

See InvalidAutoRollbackConfigException

invalid_auto_scaling_group_exception

See InvalidAutoScalingGroupException

invalid_blue_green_deployment_configuration_exception

See InvalidBlueGreenDeploymentConfigurationException

invalid_bucket_name_filter_exception

See InvalidBucketNameFilterException

invalid_compute_platform_exception

See InvalidComputePlatformException

invalid_deployed_state_filter_exception

See InvalidDeployedStateFilterException

invalid_deployment_config_name_exception

See InvalidDeploymentConfigNameException

invalid_deployment_group_name_exception

See InvalidDeploymentGroupNameException

invalid_deployment_id_exception

See InvalidDeploymentIdException

invalid_deployment_instance_type_exception

See InvalidDeploymentInstanceTypeException

invalid_deployment_status_exception

See InvalidDeploymentStatusException

invalid_deployment_style_exception

See InvalidDeploymentStyleException

invalid_deployment_target_id_exception

See InvalidDeploymentTargetIdException

invalid_deployment_wait_type_exception

See InvalidDeploymentWaitTypeException

invalid_ec2_tag_combination_exception

See InvalidEc2TagCombinationException

invalid_ec2_tag_exception

See InvalidEc2TagException

invalid_ecs_service_exception

See InvalidEcsServiceException

invalid_external_id_exception

See InvalidExternalIdException

invalid_file_exists_behavior_exception

See InvalidFileExistsBehaviorException

invalid_git_hub_account_token_exception

See InvalidGitHubAccountTokenException

invalid_git_hub_account_token_name_exception

See InvalidGitHubAccountTokenNameException

invalid_iam_session_arn_exception

See InvalidIamSessionArnException

invalid_iam_user_arn_exception

See InvalidIamUserArnException

invalid_ignore_application_stop_failures_value_exception

See InvalidIgnoreApplicationStopFailuresValueException

invalid_input_exception

See InvalidInputException

invalid_instance_name_exception

See InvalidInstanceNameException

invalid_instance_status_exception

See InvalidInstanceStatusException

invalid_instance_type_exception

See InvalidInstanceTypeException

invalid_key_prefix_filter_exception

See InvalidKeyPrefixFilterException

invalid_lifecycle_event_hook_execution_id_exception

See InvalidLifecycleEventHookExecutionIdException

invalid_lifecycle_event_hook_execution_status_exception

See InvalidLifecycleEventHookExecutionStatusException

invalid_load_balancer_info_exception

See InvalidLoadBalancerInfoException

invalid_minimum_healthy_host_value_exception

See InvalidMinimumHealthyHostValueException

invalid_next_token_exception

See InvalidNextTokenException

invalid_on_premises_tag_combination_exception

See InvalidOnPremisesTagCombinationException

invalid_operation_exception

See InvalidOperationException

invalid_registration_status_exception

See InvalidRegistrationStatusException

invalid_revision_exception

See InvalidRevisionException

invalid_role_exception

See InvalidRoleException

invalid_sort_by_exception

See InvalidSortByException

invalid_sort_order_exception

See InvalidSortOrderException

invalid_tag_exception

See InvalidTagException

invalid_tag_filter_exception

See InvalidTagFilterException

invalid_tags_to_add_exception

See InvalidTagsToAddException

invalid_target_filter_name_exception

See InvalidTargetFilterNameException

invalid_target_group_pair_exception

See InvalidTargetGroupPairException

invalid_target_instances_exception

See InvalidTargetInstancesException

invalid_time_range_exception

See InvalidTimeRangeException

invalid_traffic_routing_configuration_exception

See InvalidTrafficRoutingConfigurationException

invalid_trigger_config_exception

See InvalidTriggerConfigException

invalid_update_outdated_instances_only_value_exception

See InvalidUpdateOutdatedInstancesOnlyValueException

lifecycle_event_already_completed_exception

See LifecycleEventAlreadyCompletedException

lifecycle_hook_limit_exceeded_exception

See LifecycleHookLimitExceededException

multiple_iam_arns_provided_exception

See MultipleIamArnsProvidedException

operation_not_supported_exception

See OperationNotSupportedException

resource_arn_required_exception

See ResourceArnRequiredException

resource_validation_exception

See ResourceValidationException

revision_does_not_exist_exception

See RevisionDoesNotExistException

revision_required_exception

See RevisionRequiredException

role_required_exception

See RoleRequiredException

tag_limit_exceeded_exception

See TagLimitExceededException

tag_required_exception

See TagRequiredException

tag_set_list_limit_exceeded_exception

See TagSetListLimitExceededException

throttling_exception

See ThrottlingException

trigger_targets_limit_exceeded_exception

See TriggerTargetsLimitExceededException

unsupported_action_for_deployment_type_exception

See UnsupportedActionForDeploymentTypeException

Structs

AddTagsToOnPremisesInstancesError

AlarmsLimitExceededException

The maximum number of alarms for a deployment group (10) was exceeded.

ApplicationAlreadyExistsException

An application with the specified name with the IAM user or AWS account already

ApplicationDoesNotExistException

The application does not exist with the IAM user or AWS account.

ApplicationLimitExceededException

More applications were attempted to be created than are allowed.

ApplicationNameRequiredException

The minimum number of required application names was not specified.

ArnNotSupportedException

The specified ARN is not supported. For example, it might be an ARN for a resource

BatchGetApplicationRevisionsError

BatchGetApplicationsError

BatchGetDeploymentGroupsError

BatchGetDeploymentInstancesError

BatchGetDeploymentTargetsError

BatchGetDeploymentsError

BatchGetOnPremisesInstancesError

BatchLimitExceededException

The maximum number of names or IDs allowed for this request (100) was exceeded.

BucketNameFilterRequiredException

A bucket name is required, but was not provided.

ContinueDeploymentError

CreateApplicationError

CreateDeploymentConfigError

CreateDeploymentError

CreateDeploymentGroupError

DeleteApplicationError

DeleteDeploymentConfigError

DeleteDeploymentGroupError

DeleteGitHubAccountTokenError

DeleteResourcesByExternalIdError

DeploymentAlreadyCompletedException

The deployment is already complete.

DeploymentConfigAlreadyExistsException

A deployment configuration with the specified name with the IAM user or AWS account

DeploymentConfigDoesNotExistException

The deployment configuration does not exist with the IAM user or AWS account.

DeploymentConfigInUseException

The deployment configuration is still in use.

DeploymentConfigLimitExceededException

The deployment configurations limit was exceeded.

DeploymentConfigNameRequiredException

The deployment configuration name was not specified.

DeploymentDoesNotExistException

The deployment with the IAM user or AWS account does not exist.

DeploymentGroupAlreadyExistsException

A deployment group with the specified name with the IAM user or AWS account already

DeploymentGroupDoesNotExistException

The named deployment group with the IAM user or AWS account does not exist.

DeploymentGroupLimitExceededException

The deployment groups limit was exceeded.

DeploymentGroupNameRequiredException

The deployment group name was not specified.

DeploymentIdRequiredException

At least one deployment ID must be specified.

DeploymentIsNotInReadyStateException

The deployment does not have a status of Ready and can't continue yet.

DeploymentLimitExceededException

The number of allowed deployments was exceeded.

DeploymentNotStartedException

The specified deployment has not started.

DeploymentTargetDoesNotExistException

The provided target ID does not belong to the attempted deployment.

DeploymentTargetIdRequiredException

A deployment target ID was not provided.

DeploymentTargetListSizeExceededException

The maximum number of targets that can be associated with an Amazon ECS or AWS Lambda

DeregisterOnPremisesInstanceError

DescriptionTooLongException

The description is too long.

EcsServiceMappingLimitExceededException

The Amazon ECS service is associated with more than one deployment groups. An Amazon

GetApplicationError

GetApplicationRevisionError

GetDeploymentConfigError

GetDeploymentError

GetDeploymentGroupError

GetDeploymentInstanceError

GetDeploymentTargetError

GetOnPremisesInstanceError

GitHubAccountTokenDoesNotExistException

No GitHub account connection exists with the named specified in the call.

GitHubAccountTokenNameRequiredException

The call is missing a required GitHub account connection name.

IamArnRequiredException

No IAM ARN was included in the request. You must use an IAM session ARN or IAM user

IamSessionArnAlreadyRegisteredException

The request included an IAM session ARN that has already been used to register a

IamUserArnAlreadyRegisteredException

The specified IAM user ARN is already registered with an on-premises instance.

IamUserArnRequiredException

An IAM user ARN was not specified.

InstanceDoesNotExistException

The specified instance does not exist in the deployment group.

InstanceIdRequiredException

The instance ID was not specified.

InstanceLimitExceededException

The maximum number of allowed on-premises instances in a single call was

InstanceNameAlreadyRegisteredException

The specified on-premises instance name is already registered.

InstanceNameRequiredException

An on-premises instance name was not specified.

InstanceNotRegisteredException

The specified on-premises instance is not registered.

InvalidAlarmConfigException

The format of the alarm configuration is invalid. Possible causes include:

InvalidApplicationNameException

The application name was specified in an invalid format.

InvalidArnException

The specified ARN is not in a valid format.

InvalidAutoRollbackConfigException

The automatic rollback configuration was specified in an invalid format. For example,

InvalidAutoScalingGroupException

The Auto Scaling group was specified in an invalid format or does not exist.

InvalidBlueGreenDeploymentConfigurationException

The configuration for the blue/green deployment group was provided in an invalid

InvalidBucketNameFilterException

The bucket name either doesn't exist or was specified in an invalid format.

InvalidComputePlatformException

The computePlatform is invalid. The computePlatform should be Lambda,

InvalidDeployedStateFilterException

The deployed state filter was specified in an invalid format.

InvalidDeploymentConfigNameException

The deployment configuration name was specified in an invalid format.

InvalidDeploymentGroupNameException

The deployment group name was specified in an invalid format.

InvalidDeploymentIdException

At least one of the deployment IDs was specified in an invalid format.

InvalidDeploymentInstanceTypeException

An instance type was specified for an in-place deployment. Instance types are

InvalidDeploymentStatusException

The specified deployment status doesn't exist or cannot be determined.

InvalidDeploymentStyleException

An invalid deployment style was specified. Valid deployment types include "IN_PLACE"

InvalidDeploymentTargetIdException

The target ID provided was not valid.

InvalidDeploymentWaitTypeException

The wait type is invalid.

InvalidEc2TagCombinationException

A call was submitted that specified both Ec2TagFilters and Ec2TagSet, but only one of

InvalidEc2TagException

The tag was specified in an invalid format.

InvalidEcsServiceException

The Amazon ECS service identifier is not valid.

InvalidExternalIdException

The external ID was specified in an invalid format.

InvalidFileExistsBehaviorException

An invalid fileExistsBehavior option was specified to determine how AWS CodeDeploy

InvalidGitHubAccountTokenException

The GitHub token is not valid.

InvalidGitHubAccountTokenNameException

The format of the specified GitHub account connection name is invalid.

InvalidIamSessionArnException

The IAM session ARN was specified in an invalid format.

InvalidIamUserArnException

The IAM user ARN was specified in an invalid format.

InvalidIgnoreApplicationStopFailuresValueException

The IgnoreApplicationStopFailures value is invalid. For AWS Lambda deployments,

InvalidInputException

The input was specified in an invalid format.

InvalidInstanceNameException

The on-premises instance name was specified in an invalid format.

InvalidInstanceStatusException

The specified instance status does not exist.

InvalidInstanceTypeException

An invalid instance type was specified for instances in a blue/green deployment. Valid

InvalidKeyPrefixFilterException

The specified key prefix filter was specified in an invalid format.

InvalidLifecycleEventHookExecutionIdException

A lifecycle event hook is invalid. Review the hooks section in your

InvalidLifecycleEventHookExecutionStatusException

The result of a Lambda validation function that verifies a lifecycle event is invalid.

InvalidLoadBalancerInfoException

An invalid load balancer name, or no load balancer name, was specified.

InvalidMinimumHealthyHostValueException

The minimum healthy instance value was specified in an invalid format.

InvalidNextTokenException

The next token was specified in an invalid format.

InvalidOnPremisesTagCombinationException

A call was submitted that specified both OnPremisesTagFilters and OnPremisesTagSet,

InvalidOperationException

An invalid operation was detected.

InvalidRegistrationStatusException

The registration status was specified in an invalid format.

InvalidRevisionException

The revision was specified in an invalid format.

InvalidRoleException

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group

InvalidSortByException

The column name to sort by is either not present or was specified in an invalid

InvalidSortOrderException

The sort order was specified in an invalid format.

InvalidTagException

The tag was specified in an invalid format.

InvalidTagFilterException

The tag filter was specified in an invalid format.

InvalidTagsToAddException

The specified tags are not valid.

InvalidTargetFilterNameException

The target filter name is invalid.

InvalidTargetGroupPairException

A target group pair associated with this deployment is not valid.

InvalidTargetInstancesException

The target instance configuration is invalid. Possible causes include:

InvalidTimeRangeException

The specified time range was specified in an invalid format.

InvalidTrafficRoutingConfigurationException

The configuration that specifies how traffic is routed during a deployment is

InvalidTriggerConfigException

The trigger was specified in an invalid format.

InvalidUpdateOutdatedInstancesOnlyValueException

The UpdateOutdatedInstancesOnly value is invalid. For AWS Lambda deployments,

LifecycleEventAlreadyCompletedException

An attempt to return the status of an already completed lifecycle event

LifecycleHookLimitExceededException

The limit for lifecycle hooks was exceeded.

ListApplicationRevisionsError

ListApplicationsError

ListDeploymentConfigsError

ListDeploymentGroupsError

ListDeploymentInstancesError

ListDeploymentTargetsError

ListDeploymentsError

ListGitHubAccountTokenNamesError

ListOnPremisesInstancesError

ListTagsForResourceError

MultipleIamArnsProvidedException

Both an IAM user ARN and an IAM session ARN were included in the request. Use only one

OperationNotSupportedException

The API used does not support the deployment.

PutLifecycleEventHookExecutionStatusError

RegisterApplicationRevisionError

RegisterOnPremisesInstanceError

RemoveTagsFromOnPremisesInstancesError

ResourceArnRequiredException

The ARN of a resource is required, but was not found.

ResourceValidationException

The specified resource could not be validated.

RevisionDoesNotExistException

The named revision does not exist with the IAM user or AWS account.

RevisionRequiredException

The revision ID was not specified.

RoleRequiredException

The role ID was not specified.

SkipWaitTimeForInstanceTerminationError

StopDeploymentError

TagLimitExceededException

The maximum allowed number of tags was exceeded.

TagRequiredException

A tag was not specified.

TagResourceError

TagSetListLimitExceededException

The number of tag groups included in the tag set list exceeded the maximum allowed

ThrottlingException

An API function was called too frequently.

TriggerTargetsLimitExceededException

The maximum allowed number of triggers was exceeded.

UnsupportedActionForDeploymentTypeException

A call was submitted that is not supported for the specified deployment type.

UntagResourceError

UpdateApplicationError

UpdateDeploymentGroupError

Enums

AddTagsToOnPremisesInstancesErrorKind

BatchGetApplicationRevisionsErrorKind

BatchGetApplicationsErrorKind

BatchGetDeploymentGroupsErrorKind

BatchGetDeploymentInstancesErrorKind

BatchGetDeploymentTargetsErrorKind

BatchGetDeploymentsErrorKind

BatchGetOnPremisesInstancesErrorKind

ContinueDeploymentErrorKind

CreateApplicationErrorKind

CreateDeploymentConfigErrorKind

CreateDeploymentErrorKind

CreateDeploymentGroupErrorKind

DeleteApplicationErrorKind

DeleteDeploymentConfigErrorKind

DeleteDeploymentGroupErrorKind

DeleteGitHubAccountTokenErrorKind

DeleteResourcesByExternalIdErrorKind

DeregisterOnPremisesInstanceErrorKind

GetApplicationErrorKind

GetApplicationRevisionErrorKind

GetDeploymentConfigErrorKind

GetDeploymentErrorKind

GetDeploymentGroupErrorKind

GetDeploymentInstanceErrorKind

GetDeploymentTargetErrorKind

GetOnPremisesInstanceErrorKind

ListApplicationRevisionsErrorKind

ListApplicationsErrorKind

ListDeploymentConfigsErrorKind

ListDeploymentGroupsErrorKind

ListDeploymentInstancesErrorKind

ListDeploymentTargetsErrorKind

ListDeploymentsErrorKind

ListGitHubAccountTokenNamesErrorKind

ListOnPremisesInstancesErrorKind

ListTagsForResourceErrorKind

PutLifecycleEventHookExecutionStatusErrorKind

RegisterApplicationRevisionErrorKind

RegisterOnPremisesInstanceErrorKind

RemoveTagsFromOnPremisesInstancesErrorKind

SkipWaitTimeForInstanceTerminationErrorKind

StopDeploymentErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind

UpdateDeploymentGroupErrorKind