Struct aws_sdk_codedeploy::input::create_deployment_group_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateDeploymentGroupInput

Implementations

impl Builder

pub fn application_name(self, input: impl Into<String>) -> Self

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

pub fn set_application_name(self, input: Option<String>) -> Self

pub fn deployment_group_name(self, input: impl Into<String>) -> Self

The name of a new deployment group for the specified application.

pub fn set_deployment_group_name(self, input: Option<String>) -> Self

pub fn deployment_config_name(self, input: impl Into<String>) -> Self

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or deployment group.

For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Configurations in CodeDeploy in the AWS CodeDeploy User Guide.

pub fn set_deployment_config_name(self, input: Option<String>) -> Self

pub fn ec2_tag_filters(self, input: impl Into<Ec2TagFilter>) -> Self

pub fn set_ec2_tag_filters(self, input: Option<Vec<Ec2TagFilter>>) -> Self

pub fn on_premises_instance_tag_filters(
self,
input: impl Into<TagFilter>
) -> Self

pub fn set_on_premises_instance_tag_filters(
self,
input: Option<Vec<TagFilter>>
) -> Self

pub fn auto_scaling_groups(self, input: impl Into<String>) -> Self

pub fn set_auto_scaling_groups(self, input: Option<Vec<String>>) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

A service role Amazon Resource Name (ARN) that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn trigger_configurations(self, input: impl Into<TriggerConfig>) -> Self

pub fn set_trigger_configurations(
self,
input: Option<Vec<TriggerConfig>>
) -> Self

pub fn alarm_configuration(self, input: AlarmConfiguration) -> Self

Information to add about Amazon CloudWatch alarms when the deployment group is created.

pub fn set_alarm_configuration(self, input: Option<AlarmConfiguration>) -> Self

pub fn auto_rollback_configuration(
self,
input: AutoRollbackConfiguration
) -> Self

Configuration information for an automatic rollback that is added when a deployment group is created.

pub fn set_auto_rollback_configuration(
self,
input: Option<AutoRollbackConfiguration>
) -> Self

pub fn outdated_instances_strategy(
self,
input: OutdatedInstancesStrategy
) -> Self

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

pub fn set_outdated_instances_strategy(
self,
input: Option<OutdatedInstancesStrategy>
) -> Self

pub fn deployment_style(self, input: DeploymentStyle) -> Self

Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

pub fn set_deployment_style(self, input: Option<DeploymentStyle>) -> Self

pub fn blue_green_deployment_configuration(
self,
input: BlueGreenDeploymentConfiguration
) -> Self

Information about blue/green deployment options for a deployment group.

pub fn set_blue_green_deployment_configuration(
self,
input: Option<BlueGreenDeploymentConfiguration>
) -> Self

pub fn load_balancer_info(self, input: LoadBalancerInfo) -> Self

Information about the load balancer used in a deployment.

pub fn set_load_balancer_info(self, input: Option<LoadBalancerInfo>) -> Self

pub fn ec2_tag_set(self, input: Ec2TagSet) -> Self

Information about groups of tags applied to EC2 instances. The deployment group includes only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

pub fn set_ec2_tag_set(self, input: Option<Ec2TagSet>) -> Self

pub fn ecs_services(self, input: impl Into<EcsService>) -> Self

pub fn set_ecs_services(self, input: Option<Vec<EcsService>>) -> Self

pub fn on_premises_tag_set(self, input: OnPremisesTagSet) -> Self

Information about groups of tags applied to on-premises instances. The deployment group includes only on-premises instances identified by all of the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

pub fn set_on_premises_tag_set(self, input: Option<OnPremisesTagSet>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> Result<CreateDeploymentGroupInput, BuildError>

Consumes the builder and constructs a CreateDeploymentGroupInput