Module aws_sdk_codedeploy::model

Modules

alarm

See Alarm

alarm_configuration

See AlarmConfiguration

app_spec_content

See AppSpecContent

application_info

See ApplicationInfo

auto_rollback_configuration

See AutoRollbackConfiguration

auto_scaling_group

See AutoScalingGroup

blue_green_deployment_configuration

See BlueGreenDeploymentConfiguration

blue_instance_termination_option

See BlueInstanceTerminationOption

cloud_formation_target

See CloudFormationTarget

deployment_config_info

See DeploymentConfigInfo

deployment_group_info

See DeploymentGroupInfo

deployment_info

See DeploymentInfo

deployment_overview

See DeploymentOverview

deployment_ready_option

See DeploymentReadyOption

deployment_style

See DeploymentStyle

deployment_target

See DeploymentTarget

diagnostics

See Diagnostics

ec2_tag_filter

See Ec2TagFilter

ec2_tag_set

See Ec2TagSet

ecs_service

See EcsService

ecs_target

See EcsTarget

ecs_task_set

See EcsTaskSet

elb_info

See ElbInfo

error_information

See ErrorInformation

generic_revision_info

See GenericRevisionInfo

git_hub_location

See GitHubLocation

green_fleet_provisioning_option

See GreenFleetProvisioningOption

instance_info

See InstanceInfo

instance_summary

See InstanceSummary

instance_target

See InstanceTarget

lambda_function_info

See LambdaFunctionInfo

lambda_target

See LambdaTarget

last_deployment_info

See LastDeploymentInfo

lifecycle_event

See LifecycleEvent

load_balancer_info

See LoadBalancerInfo

minimum_healthy_hosts

See MinimumHealthyHosts

on_premises_tag_set

See OnPremisesTagSet

raw_string

See RawString

related_deployments

See RelatedDeployments

revision_info

See RevisionInfo

revision_location

See RevisionLocation

rollback_info

See RollbackInfo

s3_location

See S3Location

tag

See Tag

tag_filter

See TagFilter

target_group_info

See TargetGroupInfo

target_group_pair_info

See TargetGroupPairInfo

target_instances

See TargetInstances

time_based_canary

See TimeBasedCanary

time_based_linear

See TimeBasedLinear

time_range

See TimeRange

traffic_route

See TrafficRoute

traffic_routing_config

See TrafficRoutingConfig

trigger_config

See TriggerConfig

Structs

Alarm

Information about an alarm.

AlarmConfiguration

Information about alarms associated with the deployment group.

AppSpecContent

A revision for an AWS Lambda or Amazon ECS deployment that is a YAML-formatted or

ApplicationInfo

Information about an application.

AutoRollbackConfiguration

Information about a configuration for automatically rolling back to a previous version

AutoScalingGroup

Information about an Auto Scaling group.

BlueGreenDeploymentConfiguration

Information about blue/green deployment options for a deployment group.

BlueInstanceTerminationOption

Information about whether instances in the original environment are terminated when a

CloudFormationTarget

Information about the target to be updated by an AWS CloudFormation blue/green

DeploymentConfigInfo

Information about a deployment configuration.

DeploymentGroupInfo

Information about a deployment group.

DeploymentInfo

Information about a deployment.

DeploymentOverview

Information about the deployment status of the instances in the deployment.

DeploymentReadyOption

Information about how traffic is rerouted to instances in a replacement environment in

DeploymentStyle

Information about the type of deployment, either in-place or blue/green, you want to

DeploymentTarget

Information about the deployment target.

Diagnostics

Diagnostic information about executable scripts that are part of a deployment.

Ec2TagFilter

Information about an EC2 tag filter.

Ec2TagSet

Information about groups of EC2 instance tags.

EcsService

Contains the service and cluster names used to identify an Amazon ECS deployment's

EcsTarget

Information about the target of an Amazon ECS deployment.

EcsTaskSet

Information about a set of Amazon ECS tasks in an AWS CodeDeploy deployment. An

ElbInfo

Information about a load balancer in Elastic Load Balancing to use in a deployment.

ErrorInformation

Information about a deployment error.

GenericRevisionInfo

Information about an application revision.

GitHubLocation

Information about the location of application artifacts stored in GitHub.

GreenFleetProvisioningOption

Information about the instances that belong to the replacement environment in a

InstanceInfo

Information about an on-premises instance.

InstanceSummary

Information about an instance in a deployment.

InstanceTarget

A target Amazon EC2 or on-premises instance during a deployment that uses the

LambdaFunctionInfo

Information about a Lambda function specified in a deployment.

LambdaTarget

Information about the target AWS Lambda function during an AWS Lambda deployment.

LastDeploymentInfo

Information about the most recent attempted or successful deployment to a deployment

LifecycleEvent

Information about a deployment lifecycle event.

LoadBalancerInfo

Information about the Elastic Load Balancing load balancer or target group used in a

MinimumHealthyHosts

Information about minimum healthy instance.

OnPremisesTagSet

Information about groups of on-premises instance tags.

RawString

A revision for an AWS Lambda deployment that is a YAML-formatted or JSON-formatted

RelatedDeployments

Information about deployments related to the specified deployment.

RevisionInfo

Information about an application revision.

RevisionLocation

Information about the location of an application revision.

RollbackInfo

Information about a deployment rollback.

S3Location

Information about the location of application artifacts stored in Amazon S3.

Tag

Information about a tag.

TagFilter

Information about an on-premises instance tag filter.

TargetGroupInfo

Information about a target group in Elastic Load Balancing to use in a deployment.

TargetGroupPairInfo

Information about two target groups and how traffic is routed during an Amazon ECS

TargetInstances

Information about the instances to be used in the replacement environment in a

TimeBasedCanary

A configuration that shifts traffic from one version of a Lambda function or ECS task

TimeBasedLinear

A configuration that shifts traffic from one version of a Lambda function or ECS task

TimeRange

Information about a time range.

TrafficRoute

Information about a listener. The listener contains the path used to route traffic

TrafficRoutingConfig

The configuration that specifies how traffic is shifted from one version of a Lambda

TriggerConfig

Information about notification triggers for the deployment group.

Enums

ApplicationRevisionSortBy

AutoRollbackEvent

BundleType

ComputePlatform

DeploymentCreator

DeploymentOption

DeploymentReadyAction

DeploymentStatus

DeploymentTargetType

DeploymentType

DeploymentWaitType

Ec2TagFilterType

ErrorCode

FileExistsBehavior

GreenFleetProvisioningAction

InstanceAction

InstanceStatus

NOTE: InstanceStatus::Unknown has been renamed to ::UnknownValue.

InstanceType

LifecycleErrorCode

LifecycleEventStatus

NOTE: LifecycleEventStatus::Unknown has been renamed to ::UnknownValue.

ListStateFilterAction

MinimumHealthyHostsType

OutdatedInstancesStrategy

RegistrationStatus

RevisionLocationType

SortOrder

StopStatus

TagFilterType

TargetFilterName

TargetLabel

TargetStatus

NOTE: TargetStatus::Unknown has been renamed to ::UnknownValue.

TrafficRoutingType

TriggerEventType