Struct aws_sdk_codedeploy::model::deployment_group_info::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DeploymentGroupInfo

Implementations

impl Builder

pub fn application_name(self, input: impl Into<String>) -> Self

The application name.

pub fn set_application_name(self, input: Option<String>) -> Self

pub fn deployment_group_id(self, input: impl Into<String>) -> Self

The deployment group ID.

pub fn set_deployment_group_id(self, input: Option<String>) -> Self

pub fn deployment_group_name(self, input: impl Into<String>) -> Self

The deployment group name.

pub fn set_deployment_group_name(self, input: Option<String>) -> Self

pub fn deployment_config_name(self, input: impl Into<String>) -> Self

The deployment configuration name.

pub fn set_deployment_config_name(self, input: Option<String>) -> Self

pub fn ec2_tag_filters(self, input: impl Into<Ec2TagFilter>) -> Self

pub fn set_ec2_tag_filters(self, input: Option<Vec<Ec2TagFilter>>) -> Self

pub fn on_premises_instance_tag_filters(
self,
input: impl Into<TagFilter>
) -> Self

pub fn set_on_premises_instance_tag_filters(
self,
input: Option<Vec<TagFilter>>
) -> Self

pub fn auto_scaling_groups(self, input: impl Into<AutoScalingGroup>) -> Self

pub fn set_auto_scaling_groups(
self,
input: Option<Vec<AutoScalingGroup>>
) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

A service role Amazon Resource Name (ARN) that grants CodeDeploy permission to make calls to AWS services on your behalf. For more information, see Create a Service Role for AWS CodeDeploy in the AWS CodeDeploy User Guide.

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn target_revision(self, input: RevisionLocation) -> Self

Information about the deployment group's target revision, including type and location.

pub fn set_target_revision(self, input: Option<RevisionLocation>) -> Self

pub fn trigger_configurations(self, input: impl Into<TriggerConfig>) -> Self

pub fn set_trigger_configurations(
self,
input: Option<Vec<TriggerConfig>>
) -> Self

pub fn alarm_configuration(self, input: AlarmConfiguration) -> Self

A list of alarms associated with the deployment group.

pub fn set_alarm_configuration(self, input: Option<AlarmConfiguration>) -> Self

pub fn auto_rollback_configuration(
self,
input: AutoRollbackConfiguration
) -> Self

Information about the automatic rollback configuration associated with the deployment group.

pub fn set_auto_rollback_configuration(
self,
input: Option<AutoRollbackConfiguration>
) -> Self

pub fn deployment_style(self, input: DeploymentStyle) -> Self

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

pub fn set_deployment_style(self, input: Option<DeploymentStyle>) -> Self

pub fn outdated_instances_strategy(
self,
input: OutdatedInstancesStrategy
) -> Self

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

pub fn set_outdated_instances_strategy(
self,
input: Option<OutdatedInstancesStrategy>
) -> Self

pub fn blue_green_deployment_configuration(
self,
input: BlueGreenDeploymentConfiguration
) -> Self

Information about blue/green deployment options for a deployment group.

pub fn set_blue_green_deployment_configuration(
self,
input: Option<BlueGreenDeploymentConfiguration>
) -> Self

pub fn load_balancer_info(self, input: LoadBalancerInfo) -> Self

Information about the load balancer to use in a deployment.

pub fn set_load_balancer_info(self, input: Option<LoadBalancerInfo>) -> Self

pub fn last_successful_deployment(self, input: LastDeploymentInfo) -> Self

Information about the most recent successful deployment to the deployment group.

pub fn set_last_successful_deployment(
self,
input: Option<LastDeploymentInfo>
) -> Self

pub fn last_attempted_deployment(self, input: LastDeploymentInfo) -> Self

Information about the most recent attempted deployment to the deployment group.

pub fn set_last_attempted_deployment(
self,
input: Option<LastDeploymentInfo>
) -> Self

pub fn ec2_tag_set(self, input: Ec2TagSet) -> Self

Information about groups of tags applied to an EC2 instance. The deployment group includes only EC2 instances identified by all of the tag groups. Cannot be used in the same call as ec2TagFilters.

pub fn set_ec2_tag_set(self, input: Option<Ec2TagSet>) -> Self

pub fn on_premises_tag_set(self, input: OnPremisesTagSet) -> Self

Information about groups of tags applied to an on-premises instance. The deployment group includes only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

pub fn set_on_premises_tag_set(self, input: Option<OnPremisesTagSet>) -> Self

pub fn compute_platform(self, input: ComputePlatform) -> Self

The destination platform type for the deployment (Lambda, Server, or ECS).

pub fn set_compute_platform(self, input: Option<ComputePlatform>) -> Self

pub fn ecs_services(self, input: impl Into<EcsService>) -> Self

pub fn set_ecs_services(self, input: Option<Vec<EcsService>>) -> Self

pub fn build(self) -> DeploymentGroupInfo

Consumes the builder and constructs a DeploymentGroupInfo