Struct aws_sdk_codedeploy::model::DeploymentGroupInfo

#[non_exhaustive]
pub struct DeploymentGroupInfo {

    pub application_name: Option<String>,
    pub deployment_group_id: Option<String>,
    pub deployment_group_name: Option<String>,
    pub deployment_config_name: Option<String>,
    pub ec2_tag_filters: Option<Vec<Ec2TagFilter>>,
    pub on_premises_instance_tag_filters: Option<Vec<TagFilter>>,
    pub auto_scaling_groups: Option<Vec<AutoScalingGroup>>,
    pub service_role_arn: Option<String>,
    pub target_revision: Option<RevisionLocation>,
    pub trigger_configurations: Option<Vec<TriggerConfig>>,
    pub alarm_configuration: Option<AlarmConfiguration>,
    pub auto_rollback_configuration: Option<AutoRollbackConfiguration>,
    pub deployment_style: Option<DeploymentStyle>,
    pub outdated_instances_strategy: Option<OutdatedInstancesStrategy>,
    pub blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>,
    pub load_balancer_info: Option<LoadBalancerInfo>,
    pub last_successful_deployment: Option<LastDeploymentInfo>,
    pub last_attempted_deployment: Option<LastDeploymentInfo>,
    pub ec2_tag_set: Option<Ec2TagSet>,
    pub on_premises_tag_set: Option<OnPremisesTagSet>,
    pub compute_platform: Option<ComputePlatform>,
    pub ecs_services: Option<Vec<EcsService>>,

}

Information about a deployment group.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

application_name: Option<String>

The application name.

deployment_group_id: Option<String>

The deployment group ID.

deployment_group_name: Option<String>

The deployment group name.

deployment_config_name: Option<String>

The deployment configuration name.

ec2_tag_filters: Option<Vec<Ec2TagFilter>>

The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags.

on_premises_instance_tag_filters: Option<Vec<TagFilter>>

The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags.

auto_scaling_groups: Option<Vec<AutoScalingGroup>>

A list of associated Auto Scaling groups.

service_role_arn: Option<String>

A service role Amazon Resource Name (ARN) that grants CodeDeploy permission to make calls to AWS services on your behalf. For more information, see Create a Service Role for AWS CodeDeploy in the AWS CodeDeploy User Guide.

target_revision: Option<RevisionLocation>

Information about the deployment group's target revision, including type and location.

trigger_configurations: Option<Vec<TriggerConfig>>

Information about triggers associated with the deployment group.

alarm_configuration: Option<AlarmConfiguration>

A list of alarms associated with the deployment group.

auto_rollback_configuration: Option<AutoRollbackConfiguration>

Information about the automatic rollback configuration associated with the deployment group.

deployment_style: Option<DeploymentStyle>

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

outdated_instances_strategy: Option<OutdatedInstancesStrategy>

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>

Information about blue/green deployment options for a deployment group.

load_balancer_info: Option<LoadBalancerInfo>

Information about the load balancer to use in a deployment.

last_successful_deployment: Option<LastDeploymentInfo>

Information about the most recent successful deployment to the deployment group.

last_attempted_deployment: Option<LastDeploymentInfo>

Information about the most recent attempted deployment to the deployment group.

ec2_tag_set: Option<Ec2TagSet>

Information about groups of tags applied to an EC2 instance. The deployment group includes only EC2 instances identified by all of the tag groups. Cannot be used in the same call as ec2TagFilters.

on_premises_tag_set: Option<OnPremisesTagSet>

Information about groups of tags applied to an on-premises instance. The deployment group includes only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

compute_platform: Option<ComputePlatform>

The destination platform type for the deployment (Lambda, Server, or ECS).

ecs_services: Option<Vec<EcsService>>

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format :.

Implementations

impl DeploymentGroupInfo

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DeploymentGroupInfo