Struct aws_sdk_codedeploy::model::DeploymentInfo

#[non_exhaustive]
pub struct DeploymentInfo {

    pub application_name: Option<String>,
    pub deployment_group_name: Option<String>,
    pub deployment_config_name: Option<String>,
    pub deployment_id: Option<String>,
    pub previous_revision: Option<RevisionLocation>,
    pub revision: Option<RevisionLocation>,
    pub status: Option<DeploymentStatus>,
    pub error_information: Option<ErrorInformation>,
    pub create_time: Option<Instant>,
    pub start_time: Option<Instant>,
    pub complete_time: Option<Instant>,
    pub deployment_overview: Option<DeploymentOverview>,
    pub description: Option<String>,
    pub creator: Option<DeploymentCreator>,
    pub ignore_application_stop_failures: bool,
    pub auto_rollback_configuration: Option<AutoRollbackConfiguration>,
    pub update_outdated_instances_only: bool,
    pub rollback_info: Option<RollbackInfo>,
    pub deployment_style: Option<DeploymentStyle>,
    pub target_instances: Option<TargetInstances>,
    pub instance_termination_wait_time_started: bool,
    pub blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>,
    pub load_balancer_info: Option<LoadBalancerInfo>,
    pub additional_deployment_status_info: Option<String>,
    pub file_exists_behavior: Option<FileExistsBehavior>,
    pub deployment_status_messages: Option<Vec<String>>,
    pub compute_platform: Option<ComputePlatform>,
    pub external_id: Option<String>,
    pub related_deployments: Option<RelatedDeployments>,

}

Information about a deployment.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

application_name: Option<String>

The application name.

deployment_group_name: Option<String>

The deployment group name.

deployment_config_name: Option<String>

The deployment configuration name.

deployment_id: Option<String>

The unique ID of a deployment.

previous_revision: Option<RevisionLocation>

Information about the application revision that was deployed to the deployment group before the most recent successful deployment.

revision: Option<RevisionLocation>

Information about the location of stored application artifacts and the service from which to retrieve them.

status: Option<DeploymentStatus>

The current state of the deployment as a whole.

error_information: Option<ErrorInformation>

Information about any error associated with this deployment.

create_time: Option<Instant>

A timestamp that indicates when the deployment was created.

start_time: Option<Instant>

A timestamp that indicates when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time might be later than the complete time. This is due to differences in the clock settings of backend servers that participate in the deployment process.

complete_time: Option<Instant>

A timestamp that indicates when the deployment was complete.

deployment_overview: Option<DeploymentOverview>

A summary of the deployment status of the instances in the deployment.

description: Option<String>

A comment about the deployment.

creator: Option<DeploymentCreator>

The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Amazon EC2 Auto Scaling created the deployment.

codeDeployRollback: A rollback process created the deployment.

CodeDeployAutoUpdate: An auto-update process created the deployment when it detected outdated EC2 instances.

ignore_application_stop_failures: bool

If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

auto_rollback_configuration: Option<AutoRollbackConfiguration>

Information about the automatic rollback configuration associated with the deployment.

update_outdated_instances_only: bool

Indicates whether only instances that are not running the latest application revision are to be deployed to.

rollback_info: Option<RollbackInfo>

Information about a deployment rollback.

deployment_style: Option<DeploymentStyle>

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

target_instances: Option<TargetInstances>

Information about the instances that belong to the replacement environment in a blue/green deployment.

instance_termination_wait_time_started: bool

Indicates whether the wait period set for the termination of instances in the original environment has started. Status is 'false' if the KEEP_ALIVE option is specified. Otherwise, 'true' as soon as the termination wait period starts.

blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>

Information about blue/green deployment options for this deployment.

load_balancer_info: Option<LoadBalancerInfo>

Information about the load balancer used in the deployment.

additional_deployment_status_info: Option<String>

Provides information about the results of a deployment, such as whether instances in the original environment in a blue/green deployment were not terminated.

file_exists_behavior: Option<FileExistsBehavior>

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

deployment_status_messages: Option<Vec<String>>

Messages that contain information about the status of a deployment.

compute_platform: Option<ComputePlatform>

The destination platform type for the deployment (Lambda, Server, or ECS).

external_id: Option<String>

The unique ID for an external resource (for example, a CloudFormation stack ID) that is linked to this deployment.

related_deployments: Option<RelatedDeployments>

Information about deployments related to the specified deployment.

Implementations

impl DeploymentInfo

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DeploymentInfo