Module aws_sdk_codedeploy::operation

Structs

AddTagsToOnPremisesInstances

Adds tags to on-premises instances.

BatchGetApplicationRevisions

Gets information about one or more application revisions. The maximum number of

BatchGetApplications

Gets information about one or more applications. The maximum number of applications

BatchGetDeploymentGroups

Gets information about one or more deployment groups.

BatchGetDeploymentInstances

BatchGetDeploymentTargets

Returns an array of one or more targets associated with a deployment. This method

BatchGetDeployments

Gets information about one or more deployments. The maximum number of deployments that

BatchGetOnPremisesInstances

Gets information about one or more on-premises instances. The maximum number of

ContinueDeployment

For a blue/green deployment, starts the process of rerouting traffic from instances in

CreateApplication

Creates an application.

CreateDeployment

Deploys an application revision through the specified deployment group.

CreateDeploymentConfig

Creates a deployment configuration.

CreateDeploymentGroup

Creates a deployment group to which application revisions are deployed.

DeleteApplication

Deletes an application.

DeleteDeploymentConfig

Deletes a deployment configuration.

DeleteDeploymentGroup

Deletes a deployment group.

DeleteGitHubAccountToken

Deletes a GitHub account connection.

DeleteResourcesByExternalId

Deletes resources linked to an external ID.

DeregisterOnPremisesInstance

Deregisters an on-premises instance.

GetApplication

Gets information about an application.

GetApplicationRevision

Gets information about an application revision.

GetDeployment

Gets information about a deployment.

GetDeploymentConfig

Gets information about a deployment configuration.

GetDeploymentGroup

Gets information about a deployment group.

GetDeploymentInstance

Gets information about an instance as part of a deployment.

GetDeploymentTarget

Returns information about a deployment target.

GetOnPremisesInstance

Gets information about an on-premises instance.

ListApplicationRevisions

Lists information about revisions for an application.

ListApplications

Lists the applications registered with the IAM user or AWS account.

ListDeploymentConfigs

Lists the deployment configurations with the IAM user or AWS account.

ListDeploymentGroups

Lists the deployment groups for an application registered with the IAM user or AWS

ListDeploymentInstances

ListDeploymentTargets

Returns an array of target IDs that are associated a deployment.

ListDeployments

Lists the deployments in a deployment group for an application registered with the IAM

ListGitHubAccountTokenNames

Lists the names of stored connections to GitHub accounts.

ListOnPremisesInstances

Gets a list of names for one or more on-premises instances.

ListTagsForResource

Returns a list of tags for the resource identified by a specified Amazon Resource

PutLifecycleEventHookExecutionStatus

Sets the result of a Lambda validation function. The function validates lifecycle

RegisterApplicationRevision

Registers with AWS CodeDeploy a revision for the specified application.

RegisterOnPremisesInstance

Registers an on-premises instance.

RemoveTagsFromOnPremisesInstances

Removes one or more tags from one or more on-premises instances.

SkipWaitTimeForInstanceTermination

In a blue/green deployment, overrides any specified wait time and starts terminating

StopDeployment

Attempts to stop an ongoing deployment.

TagResource

Associates the list of tags in the input Tags parameter with the

UntagResource

Disassociates a resource from a list of tags. The resource is identified by the

UpdateApplication

Changes the name of an application.

UpdateDeploymentGroup

Changes information about a deployment group.